requirements.txt
imageio==2.6.1
numpy>=1.17.2
Pillow==6.2.0
torch~=1.4.0
torchvision~=0.5.0
tqdm~=4.43.0
scipy==1.3.3
